Fix code remarks
