+++ modules/example/src/main/resources/application.conf
@@ -2,5 +2,11 @@ proxy {
   host = "0.0.0.0"
   port = 8080
   backend = "http://0.0.0.0:9000"
-  tracer = "0.0.0.0:9411"
 }
+
+backend {
+  host = "0.0.0.0"
+  port = 9000
+}
+
+tracer = "0.0.0.0:9411"
