--- modules/example/src/main/scala/zio/telemetry/example/backend/config/package.scala
@@ -1,7 +1,0 @@
-package zio.telemetry.example.backend
-
-import zio.RIO
-
-package object config extends Configuration.Service[Configuration] {
-  val load: RIO[Configuration, Config] = RIO.accessM(_.config.load)
-}
